requirements.txt
six
requests
beautifulsoup4
python-dateutil
feedparser
